[Chore][Cleanup] merge remaining fragment files: mhc pair, thin norm benches, elementwise strays
- tests/ops/test_mhc{_pre,_post}.py -> test_mhc.py (identical
  _cosine_compare deduped); benchmarks bench_mhc{_pre,_post}.py ->
  bench_mhc.py.
- benchmarks: bench_{rms,layer,fused_add_rms,fused_add_layer}_norm.py
  -> bench_norm.py with per-op prefixed module constants.
- tileops/ops/elementwise: fused_gated.py -> activations.py (gated
  activations); predicates.py -> comparison.py (isnan/isinf/isfinite
  follow PyTorch's comparison-ops taxonomy).
- manifest source.op/test/bench pointers retargeted accordingly.

Test/bench node names and params identical; only file paths moved.
Validator passes.
