--- benchmarks/ops/bench_layer_norm.py
@@ -1,44 +1,0 @@
-import pytest
-import torch
-import torch.nn.functional as F
-
-from benchmarks.benchmark_base import BenchmarkReport, ManifestBenchmark
-from tileops.manifest import load_workloads
-from tileops.ops.norm.layer_norm import LayerNormFwdOp
-from workloads.normalization import LayerNormTest
-
-_OP_NAME = "LayerNormFwdOp"
-
-
-def _manifest_params():
-    params = []
-    for w in load_workloads(_OP_NAME):
-        m, n = w["x_shape"]
-        label = w.get("label", f"{m}x{n}")
-        for dtype_str in w["dtypes"]:
-            dtype = getattr(torch, dtype_str)
-            params.append(pytest.param(m, n, dtype, True,
-                                       id=f"{label}-{dtype_str}"))
-    return params
-
-
-@pytest.mark.parametrize("m, n, dtype, tune", _manifest_params())
-def test_layer_norm_bench(m: int, n: int, dtype: torch.dtype, tune: bool) -> None:
-    test = LayerNormTest(m, n, dtype)
-    inputs = test.gen_inputs()
-
-    op = LayerNormFwdOp(normalized_shape=(n,), dtype=dtype, tune=tune)
-    bm = ManifestBenchmark(_OP_NAME, op, test)
-    result = bm.profile(op, *inputs)
-    BenchmarkReport.record(op, locals(), result, tag="tileops")
-
-    # Baseline uses torch.nn.functional.layer_norm
-    def baseline_fn(x, weight, bias):
-        return F.layer_norm(x, (n,), weight=weight, bias=bias, eps=1e-5)
-
-    result_bl = bm.profile(baseline_fn, *inputs)
-    BenchmarkReport.record(op, locals(), result_bl, tag="torch")
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-vvs"])
