+++ tileops/manifest/elementwise_fused_gated.yaml
@@ -46,7 +46,7 @@ SiluAndMulFwdOp:
     kernel: tileops/kernels/elementwise.py
     kernel_map:
       silu_and_mul: SiluAndMulFwdKernel
-    op: tileops/ops/elementwise/fused_gated.py
+    op: tileops/ops/elementwise/activations.py
     test: tests/ops/test_fused_gated.py
     bench: benchmarks/ops/bench_binary_elementwise.py
     bench_manifest_driven: false
@@ -87,7 +87,7 @@ GeluAndMulFwdOp:
     kernel: tileops/kernels/elementwise.py
     kernel_map:
       gelu_and_mul: GeluAndMulFwdKernel
-    op: tileops/ops/elementwise/fused_gated.py
+    op: tileops/ops/elementwise/activations.py
     test: tests/ops/test_fused_gated.py
     bench: benchmarks/ops/bench_binary_elementwise.py
     bench_manifest_driven: false
@@ -129,7 +129,7 @@ GeluTanhAndMulFwdOp:
     kernel: tileops/kernels/elementwise.py
     kernel_map:
       gelu_tanh_and_mul: GeluTanhAndMulFwdKernel
-    op: tileops/ops/elementwise/fused_gated.py
+    op: tileops/ops/elementwise/activations.py
     test: tests/ops/test_fused_gated.py
     bench: benchmarks/ops/bench_binary_elementwise.py
     bench_manifest_driven: false
