+++ tileops/ops/elementwise/activations.py
@@ -6,7 +6,9 @@ import torch
 
 from tileops.kernels.elementwise import (
     EluFwdKernel,
+    GeluAndMulFwdKernel,
     GeluFwdKernel,
+    GeluTanhAndMulFwdKernel,
     GeluTanhFwdKernel,
     HardsigmoidFwdKernel,
     HardswishFwdKernel,
@@ -16,6 +18,7 @@ from tileops.kernels.elementwise import (
     ReluFwdKernel,
     SeluFwdKernel,
     SigmoidFwdKernel,
+    SiluAndMulFwdKernel,
     SiluFwdKernel,
     SoftplusFwdKernel,
     TanhFwdKernel,
@@ -23,6 +26,7 @@ from tileops.kernels.elementwise import (
 from tileops.kernels.kernel_base import Kernel
 
 from ._base import (
+    FusedGatedOp,
     UnaryOp,
     _GeluApproximateBase,
     _ParametricActivationOp,
@@ -320,3 +324,25 @@ class SoftplusFwdOp(_ParametricActivationOp):
     @property
     def default_kernel_map(self):
         return {"softplus": SoftplusFwdKernel}
+
+
+class SiluAndMulFwdOp(FusedGatedOp):
+    """SiLU-and-Mul: y = silu(gate) * value."""
+
+    _op_name = "silu_and_mul"
+    kernel_cls = SiluAndMulFwdKernel
+
+
+class GeluAndMulFwdOp(FusedGatedOp):
+    """GELU-and-Mul: y = gelu(gate) * value (exact GELU)."""
+
+    _op_name = "gelu_and_mul"
+    kernel_cls = GeluAndMulFwdKernel
+
+
+class GeluTanhAndMulFwdOp(FusedGatedOp):
+    """GELU-Tanh-and-Mul: y = gelu_tanh(gate) * value (tanh approximation)."""
+
+    _op_name = "gelu_tanh_and_mul"
+    kernel_cls = GeluTanhAndMulFwdKernel
+    FLOPS_PER_ELEM = 10
