+++ tileops/ops/elementwise/comparison.py
@@ -1,5 +1,7 @@
 """Element-wise comparison ops (output bool)."""
 
+import torch
+
 from tileops.kernels.elementwise import (
     EqBoolStorageFwdKernel,
     EqFwdKernel,
@@ -7,6 +9,9 @@ from tileops.kernels.elementwise import (
     GeFwdKernel,
     GtBoolStorageFwdKernel,
     GtFwdKernel,
+    IsfiniteFwdKernel,
+    IsinfFwdKernel,
+    IsnanFwdKernel,
     LeBoolStorageFwdKernel,
     LeFwdKernel,
     LtBoolStorageFwdKernel,
@@ -15,7 +20,13 @@ from tileops.kernels.elementwise import (
     NeFwdKernel,
 )
 
-from ._base import _BoolOutputBinaryOp
+from ._base import (
+    _PREDICATE_FALLBACK_DTYPES,
+    _BoolOutputBinaryOp,
+    _int_all_false,
+    _int_all_true,
+    _IntIdentityUnaryOp,
+)
 
 
 class EqFwdOp(_BoolOutputBinaryOp):
@@ -64,3 +75,45 @@ class LeFwdOp(_BoolOutputBinaryOp):
     _op_name = "le"
     kernel_cls = LeFwdKernel
     bool_storage_kernel_cls = LeBoolStorageFwdKernel
+
+
+class IsnanFwdOp(_IntIdentityUnaryOp):
+    """Element-wise isnan with bool output.
+
+    Always False on integer / bool input (no NaN representation in those
+    dtypes).
+    """
+
+    _op_name = "isnan"
+    kernel_cls = IsnanFwdKernel
+    _int_handler = staticmethod(_int_all_false)
+    _int_output_dtype = torch.bool
+    _fallback_dtypes = _PREDICATE_FALLBACK_DTYPES
+
+
+class IsinfFwdOp(_IntIdentityUnaryOp):
+    """Element-wise isinf with bool output.
+
+    Always False on integer / bool input (no Inf representation in those
+    dtypes).
+    """
+
+    _op_name = "isinf"
+    kernel_cls = IsinfFwdKernel
+    _int_handler = staticmethod(_int_all_false)
+    _int_output_dtype = torch.bool
+    _fallback_dtypes = _PREDICATE_FALLBACK_DTYPES
+
+
+class IsfiniteFwdOp(_IntIdentityUnaryOp):
+    """Element-wise isfinite with bool output.
+
+    Always True on integer / bool input (every value in those dtypes is
+    finite).
+    """
+
+    _op_name = "isfinite"
+    kernel_cls = IsfiniteFwdKernel
+    _int_handler = staticmethod(_int_all_true)
+    _int_output_dtype = torch.bool
+    _fallback_dtypes = _PREDICATE_FALLBACK_DTYPES
